Create Range() with For loop
An interesting control flow challenge. Whereby I have used a combination of range() in-built function and For loop and an if-else statement to check if a no falls within a certain range or not.
A simpler method would have been to use a single if(with an 'and' within) else statement, but this one was more satisfying.